refactor(tauri): extract the Cowork commands into cowork_commands.rs
Unit 11d of the lib.rs split. 1579 lines move verbatim -- the eleven
Cowork invoke commands with their non-Windows stubs, the pure decision
helpers, `iso_now`/`is_leap` (whose only callers are in the cluster) and
all four test modules. 23 lines differ from their originals and every
one of them is a `pub(crate)` widening so `generate_handler!` and the
one non-command caller in `lib.rs` can still name them.

The import block is where a move this size goes wrong, and the failure
is asymmetric. Five of the sibling modules the cluster calls --
`cowork_atomic_json`, `cowork_workspace_scan`, `cowork_installer`,
`firewall`, `cowork_meta` -- have `mod` declarations in `lib.rs` that are
themselves `#[cfg(target_os = "windows")]`. Off Windows those modules do
not exist at all, so an ungated `use crate::firewall;` here is an
unconditional E0432 on macOS and Linux while compiling clean on a
Windows dev box. Unit 11c shipped exactly that and only CI caught it.
`token_store` joins the gated group despite being an ungated module,
because every call site here is Windows-gated. `single_flight` stays
ungated: `SUBNET_PROBE_FLIGHT` sits outside any cfg.

`lib.rs`'s own `PathBuf` import narrows from
`#[cfg(any(test, target_os = "windows"))]` to `#[cfg(test)]`. This
cluster was its last non-test Windows consumer, and leaving the wider
gate would warn on a Windows RELEASE build alone -- which neither
`cargo test` nor any CI leg can see.

The two guards that read the cluster as source text move off a hardcoded
`src-tauri/src/lib.rs` and onto `rustSourceDefining`, which finds the
module by a construct that IS the subject. A hardcoded path does not go
red across this split, it goes quiet.

Verified: `cargo test --lib` 285 passed, with all 18 moved tests
reporting under `cowork_commands::`. `cargo check --all-targets` clean.
